feat: adiciona async e Promise a function da rota upload-ceap

+++ backend/prod/app.js
@@ -16,27 +16,37 @@ app.get("/deputados", (req, res) => {
     const { uf } = req.query;
     //pesquisa todos os deputados com a uf indicada
 });
-app.post("/upload-ceap", upload.single("ceapFile"), (req, res) => {
+app.post("/upload-ceap", upload.single("ceapFile"), async (req, res) => {
     if (!req.file) {
         res.status(400).json({ error: "Nenhum arquivo CSV enviado." });
     }
     else {
         const results = [];
         const filePath = req.file.path;
-        fs_1.default.createReadStream(filePath)
-            .pipe((0, csv_parser_1.default)({ separator: ';' }))
-            .on("data", (data) => {
-            if (data.sgUF !== "NA") {
-                results.push(data);
+        try {
+            await new Promise((resolve, reject) => {
+                fs_1.default.createReadStream(filePath)
+                    .pipe((0, csv_parser_1.default)({ separator: ';' }))
+                    .on("data", (data) => {
+                    if (data.sgUF && data.sgUF !== "NA") {
+                        results.push(data);
+                    }
+                })
+                    .on("end", () => {
+                    fs_1.default.unlinkSync(filePath);
+                    resolve();
+                })
+                    .on("error", (error) => {
+                    reject(error);
+                });
+            });
+        }
+        catch (error) {
+            if (fs_1.default.existsSync(filePath)) {
+                fs_1.default.unlinkSync(filePath);
             }
-        })
-            .on("end", () => {
-            fs_1.default.unlinkSync(filePath);
-            res.json({ data: results });
-        })
-            .on("error", (error) => {
-            res.status(500).json({ error: `Erro ao processar o CSV: ${error.message}` });
-        });
+            res.status(500).json({ error: `Erro ao processar o CSV` });
+        }
     }
 });
 app.listen(port, () => {
